+++ src/python_bioinformagicks/plotting/__init__.py
@@ -1,5 +1,7 @@
 from ._split_plot_embedding import split_plot_embedding 
+from ._plot_grouped_proportions import plot_grouped_proportions
 
 __all__ = [
 	"split_plot_embedding",
+	"plot_grouped_proportions",
 ]