+++ src/python_bioinformagicks/tools/__init__.py
@@ -2,7 +2,6 @@ from ._tf_idf_markers import tf_idf_markers
 from ._subset_by_coordinates import subset_by_coordinates
 from ._subset_by_geosketching import subset_by_geosketching
 from ._calc_jasmine_score import calc_jasmine_score
-from ._in_ignore_list import in_ignore_list
 from ._scale_by_group import scale_by_group
 from ._do_gprofiler_analysis import do_gprofiler_analysis