+++ src/python_bioinformagicks/utilities/__init__.py
@@ -1,0 +1,9 @@
+from ._in_ignore_list import in_ignore_list 
+from ._make_combined_categorical_column import make_combined_categorical_column
+from ._get_proportions import get_proportions
+
+__all__ = [
+	"in_ignore_list",
+	"make_combined_categorical_column",
+	"get_proportions"
+]