+++ HISTORY.rst
@@ -2,6 +2,16 @@
 History
 =======
 
+v0.3.0 (2024-12-08)
+--------------------
+
+* Added bim.util submodule
+* Moved bim.tl.in_ignore_list to bim.util
+* Added bim.pl.plot_grouped_proportions (stacked barplot)
+* Added bim.util.get_proportions
+* Added bim.util.make_combined_categorical_column
+
+
 v0.2.0 (2024-12-18)
 -------------------